# server_source/reports/openapi_manual_export.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Murim Inn Rebuild Server",
    "version": "0.4.0"
  },
  "paths": {
...
    "/admin/catalog/iap-products": {
      "get": {
        "tags": [
          "AdminCatalog"
        ],
        "summary": "Get Iap Products",
        "operationId": "get_iap_products_admin_catalog_iap_products_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "AdminCatalog"
        ],
        "summary": "Upsert Iap Product",
        "operationId": "upsert_iap_product_admin_catalog_iap_products_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IapProductUpsertReq"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/catalog/tiers": {
      "get": {
        "tags": [
          "AdminCatalog"
        ],
        "summary": "Get Tiers",
        "operationId": "get_tiers_admin_catalog_tiers_get",
        "parameters": [
          {
            "name": "seasonId",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Seasonid"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "AdminCatalog"
        ],
        "summary": "Post Tier",
        "operationId": "post_tier_admin_catalog_tiers_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TierUpsertReq"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/catalog/mbti": {
      "get": {
        "tags": [
          "AdminCatalog"
        ],
        "summary": "Get Mbti",
        "operationId": "get_mbti_admin_catalog_mbti_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "AdminCatalog"
        ],
        "summary": "Post Mbti",
        "operationId": "post_mbti_admin_catalog_mbti_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MbtiUpsertReq"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/catalog/items": {
      "get": {
        "tags": [
          "AdminCatalog"
        ],
        "summary": "Get Items",
        "operationId": "get_items_admin_catalog_items_get",
        "parameters": [
          {
            "name": "seasonId",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Seasonid"
            }
          },
          {
            "name": "tierId",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tierid"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "AdminCatalog"
        ],
        "summary": "Post Item",
        "operationId": "post_item_admin_catalog_items_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ItemUpsertReq"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/catalog/promotions": {
      "get": {
        "tags": [
          "AdminCatalog"
        ],
        "summary": "Get Promotions",
        "operationId": "get_promotions_admin_catalog_promotions_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "AdminCatalog"
        ],
        "summary": "Post Promotions",
        "operationId": "post_promotions_admin_catalog_promotions_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PromotionNodeUpsertReq"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "IapProductUpsertReq": {
        "properties": {
          "productId": {
            "type": "string",
            "title": "Productid"
          },
          "currency": {
            "type": "string",
            "pattern": "^(gold|gems|exp)$",
            "title": "Currency"
          },
          "amount": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Amount"
          }
        },
        "type": "object",
        "required": [
          "productId",
          "currency",
          "amount"
        ],
        "title": "IapProductUpsertReq"
      },
      "ItemUpsertReq": {
        "properties": {
          "itemId": {
            "type": "string",
            "title": "Itemid"
          },
          "seasonId": {
            "type": "string",
            "title": "Seasonid"
          },
          "tierId": {
            "type": "string",
            "title": "Tierid"
          },
          "slot": {
            "type": "string",
            "title": "Slot"
          },
          "atkMul": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Atkmul"
          },
          "hpMul": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Hpmul"
          },
          "defMul": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Defmul"
          },
          "skillMul": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Skillmul"
          }
        },
        "type": "object",
        "required": [
          "itemId",
          "seasonId",
          "tierId",
          "slot",
          "atkMul",
          "hpMul",
          "defMul",
          "skillMul"
        ],
        "title": "ItemUpsertReq"
      },
...
      "MbtiUpsertReq": {
        "properties": {
          "mbti": {
            "type": "string",
            "title": "Mbti"
          },
          "atkMul": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Atkmul"
          },
          "hpMul": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Hpmul"
          },
          "defMul": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Defmul"
          },
          "goldMul": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Goldmul"
          },
          "expMul": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Expmul"
          }
        },
        "type": "object",
        "required": [
          "mbti",
          "atkMul",
          "hpMul",
          "defMul",
          "goldMul",
          "expMul"
        ],
        "title": "MbtiUpsertReq"
      },
...
      "PromotionNodeUpsertReq": {
        "properties": {
          "nodeId": {
            "type": "string",
            "title": "Nodeid"
          },
          "parentNodeId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parentnodeid"
          },
          "jobId": {
            "type": "string",
            "title": "Jobid"
          },
          "choiceGroup": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Choicegroup"
          },
          "promotionMultiplier": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Promotionmultiplier"
          },
          "statBonus": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Statbonus"
          },
          "skillUnlock": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Skillunlock"
          }
        },
        "type": "object",
        "required": [
          "nodeId",
          "jobId",
          "promotionMultiplier"
        ],
        "title": "PromotionNodeUpsertReq"
      },
...
      "TierUpsertReq": {
        "properties": {
          "seasonId": {
            "type": "string",
            "title": "Seasonid"
          },
          "tierId": {
            "type": "string",
            "title": "Tierid"
          },
          "multiplier": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Multiplier"
          }
        },
        "type": "object",
        "required": [
          "seasonId",
          "tierId",
          "multiplier"
        ],
        "title": "TierUpsertReq"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}